backend/requirements.txt
flask==2.3.2
flask-cors==3.0.10
flask-session==0.5.0
nltk==3.8.1
numpy==1.24.3
spacy==3.7.0
scikit-learn==1.3.1
python-dotenv==1.0.0
PyPDF2==3.0.1
regex==2023.8.8
pandas==2.1.0
tqdm==4.66.1
matplotlib==3.8.0
python-Levenshtein==0.22.0
fuzzywuzzy==0.18.0
sentence-transformers==2.2.2